+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 - Fix support for `postcss-load-config` [#102](https://github.com/csstools/postcss-font-magician/pull/102)
 - Fix repo and homepage urls [#101](https://github.com/csstools/postcss-font-magician/pull/101)
 - Fix typo in readme [#87](https://github.com/csstools/postcss-font-magician/pull/87)
-- Fix source maps
+- Fix source maps [#109](https://github.com/csstools/postcss-font-magician/issues/109)
 - Fix output path for `async` option,this is now relative to the process.
+- Fix order of CSS, now the `@font-face` will be inserted after `@charset`, `@import` and similar statements. [#107](https://github.com/csstools/postcss-font-magician/issues/107)
 
 
 ## 3.0.0 (2020-10-03)